use shared provider shell
