+++ packages/core-apps/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./http";
 export * from "./logger";
+export * from "./provider";
