+++ packages/core-apps/tsup.config.ts
@@ -3,4 +3,7 @@ import { packageOptions, defineConfig } from "@trigger.dev/tsup";
 export default defineConfig({
   ...packageOptions,
   config: "tsconfig.build.json",
+  banner: {
+    js: "import { createRequire } from 'module';const require = createRequire(import.meta.url);",
+  },
 });
